// src/ipl-frontend/src/app/models/match.ts
export class Match {

    id: number;
    city: String;
    date: String;
    playerOfMatch: String;
    venue: String;
    neutralVenue:String;
    team1:String;
    team2:String;
    tossWinner: String;
    tossDecision: String;
    matchWinner: String;
    result: String;
    resultMargin: String;
    eliminator:String; //superover- y
    method: String;
    umpire1: String;
    umpire2: String;
    teamName: String;
    opposition: String;
}